도메인명 변경 및 비밀번호 변경 기능 개발 완료
