+++ src/main/java/com/innovation/assignment/customer/domain/vo/BirthDate.java
@@ -1,0 +1,21 @@
+package com.innovation.assignment.customer.domain.vo;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE, force = true)
+public class BirthDate {
+
+    private final String birthDate;
+
+    private BirthDate(String birthDate) {
+        this.birthDate = birthDate;
+    }
+
+    public static BirthDate of(String birthDate) {
+
+        //생년월일 형식 검증
+        //생년월일이 미래일 경우 예외 발생
+        return new BirthDate(birthDate);
+    }
+}
