+++ src/main/java/com/innovation/assignment/customer/infrastructure/dto/GetCustomerResponseDto.java
@@ -1,0 +1,19 @@
+package com.innovation.assignment.customer.infrastructure.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class GetCustomerResponseDto {
+
+    private Long id;
+    private String email;
+    private String password;
+    private String birthDate;
+    private String phone;
+    private String address;
+    private String addressDetail;
+}
