+++ src/main/java/com/innovation/assignment/customer/presentation/dto/ChangePasswordRequestDto.java
@@ -1,0 +1,8 @@
+package com.innovation.assignment.customer.presentation.dto;
+
+public record ChangePasswordRequestDto(
+        Long customerId,
+        String newPassword,
+        String confirmNewPassword
+) {
+}
