+++ src/main/java/com/innovation/assignment/customer/presentation/dto/SearchCustomerByEmailRequestDto.java
@@ -1,0 +1,6 @@
+package com.innovation.assignment.customer.presentation.dto;
+
+public record SearchCustomerByEmailRequestDto(
+        String email
+) {
+}
