+++ src/main/java/com/innovation/assignment/customer/presentation/dto/SearchCustomerByPhoneRequestDto.java
@@ -1,0 +1,6 @@
+package com.innovation.assignment.customer.presentation.dto;
+
+public record SearchCustomerByPhoneRequestDto(
+        String phone
+) {
+}
