+++ src/main/java/com/innovation/assignment/exception/exception/customer/EmptyCustomerListException.java
@@ -1,0 +1,12 @@
+package com.innovation.assignment.exception.exception.customer;
+
+import com.innovation.assignment.exception.ApplicationException;
+import com.innovation.assignment.exception.ErrorCode;
+import org.springframework.http.HttpStatus;
+
+public class EmptyCustomerListException extends ApplicationException {
+
+    public EmptyCustomerListException() {
+        super(HttpStatus.NO_CONTENT, ErrorCode.EMPTY_CUSTOMER_LIST, ErrorCode.EMPTY_CUSTOMER_LIST.getMessage());
+    }
+}
