--- src/main/java/com/innovation/assignment/exception/exception/member/DuplicateEmailException.java
@@ -1,12 +1,0 @@
-package com.innovation.assignment.exception.exception.member;
-
-import com.innovation.assignment.exception.ApplicationException;
-import com.innovation.assignment.exception.ErrorCode;
-import org.springframework.http.HttpStatus;
-
-public class DuplicateEmailException extends ApplicationException {
-
-    public DuplicateEmailException() {
-        super(HttpStatus.CONFLICT, ErrorCode.DUPLICATE_EMAIL, ErrorCode.DUPLICATE_EMAIL.getMessage());
-    }
-}
