--- src/main/java/com/innovation/assignment/member/domain/vo/Email.java
@@ -1,25 +1,0 @@
-package com.innovation.assignment.member.domain.vo;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE, force = true)
-public class Email {
-
-    private final String email;
-
-    private Email(String email) {
-        this.email = email;
-    }
-
-    public static Email of(String email) {
-
-        if (email == null || email.isBlank()) {
-            throw new IllegalArgumentException("이메일을 입력하세요.");
-        }
-        if (!email.matches("^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$")) {
-            throw new IllegalArgumentException("이메일 형식이 올바르지 않습니다.");
-        }
-        return new Email(email);
-    }
-}
