--- src/main/java/com/innovation/assignment/member/infrastructure/dto/GetMembersResponseDto.java
@@ -1,17 +1,0 @@
-package com.innovation.assignment.member.infrastructure.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class GetMembersResponseDto {
-
-    private Long id;
-    private String email;
-    private String password;
-    private String birthDate;
-    private String phone;
-    private String address;
-    private String addressDetail;
-}
