--- src/main/java/com/innovation/assignment/member/presentation/dto/SearchMemberByPhoneRequestDto.java
@@ -1,6 +1,0 @@
-package com.innovation.assignment.member.presentation.dto;
-
-public record SearchMemberByPhoneRequestDto(
-        String phone
-) {
-}
