+++ src/main/resources/application.yml
@@ -5,6 +5,8 @@ spring:
     driver-class-name: org.h2.Driver
     hikari:
       maximum-pool-size: 50
+  autoconfigure:
+    exclude: org.springframework.boot.autoconfigure.security.servlet.UserDetailsServiceAutoConfiguration
 
   jpa:
     open-in-view: false
